feat(planung): vereinheitlichte /planung-Seite mit Tabs Betreuungsplan | Dienstplan | Vertretung (#1886)
- die drei Seiten sind 1:1 als einbettbare Views extrahiert
  (betreuungsplan-view, vertretungsplan-view, dienstplan-view); kein
  visueller Neubau
- /timetables, /vertretungsplan und /staff/dienstplan leiten mit allen
  Query-Params auf /planung?tab=… weiter (Deep-Links bleiben gültig)
- Sidebar/Accordion/Breadcrumbs/Mobile-Nav auf /planung umgestellt
- Help-Guide-Schritte auf die Tab-Navigation umgeschrieben und das
  Änderungsprotokoll (Verlauf) dokumentiert
