--- frontend/src/app/[tenant]/(protected)/timetables/loading.tsx
@@ -1,12 +1,0 @@
-"use client";
-
-import { TimetablePageSkeleton } from "./page-skeleton";
-
-/**
- * Route-level loading UI: renders the same skeleton the page shows while its
- * data loads, so navigation shows one continuous skeleton instead of the
- * generic group-level Loading followed by the page skeleton.
- */
-export default function TimetablesLoading() {
-  return <TimetablePageSkeleton />;
-}
